requirements.txt
Django==3.2.5
psycopg2-binary>=2.8
petl==1.7.4
django-bootstrap4==21.1
python-dateutil==2.8.2
aiocache==0.11.1
aiohttp==3.8.1
